+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2023-04-19
+### Fixed
+- Fix an bug where the content zoom out become out of boundary when movement type is Clamped.
+- Fix PinchableScrollRect Editor to inherit ScrollRect Editor.
+
 ## [1.0.5] - 2022-12-22
 ### Fixed
 - Fix an incorrect zooming behaviour on ScreenSpace-Camera render mode.